requirements.txt
streamlit>=1.36.0
python-dotenv>=1.0.1
openai>=1.35.7
langchain>=0.2.6
langchain-openai>=0.1.7
